+++ config_file/__init__.py
@@ -1,5 +1,10 @@
-"""
-config_file: Manage your configuration files.
-"""
+from config_file.config_file import ConfigFile
+from config_file.parsers.base_parser import BaseParser, ParsingError
 
 __version__ = "0.2.0"
+
+__all__ = [
+    "ConfigFile",
+    "ParsingError",
+    "BaseParser",
+]